+++ fortnitepy/__init__.py
@@ -4,7 +4,7 @@
 """
 MIT License
 
-Copyright (c) 2019 Terbau
+Copyright (c) 2019-2020 Terbau
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -25,10 +25,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-__version__ = '1.4.0'
+__version__ = '1.5.0'
+import asyncio
+from sys import platform
 
-from .client import (Client, get_event_loop, run_multiple, start_multiple,
-                     close_multiple)
+from .client import Client, run_multiple, start_multiple, close_multiple
 from .auth import (EmailAndPasswordAuth, ExchangeCodeAuth, DeviceAuth,
                    AdvancedAuth)
 from .friend import Friend, PendingFriend
@@ -45,4 +46,5 @@ from .news import BattleRoyaleNewsPost
 from .playlist import Playlist
 
 # temporary fix for python 3.8
-get_event_loop()
+if platform == 'win32':
+    asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())