+++ fortnitepy/client.py
@@ -3,7 +3,7 @@
 """
 MIT License
 
-Copyright (c) 2019 Terbau
+Copyright (c) 2019-2020 Terbau
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -30,7 +30,6 @@ import sys
 import signal
 import logging
 import json
-import selectors
 
 from bs4 import BeautifulSoup
 from OpenSSL.SSL import SysCallError
@@ -99,33 +98,11 @@ def _cleanup_loop(loop: asyncio.AbstractEventLoop) -> None:
         loop.close()
 
 
-# Python 3.8 workaround for compatibility with aioxmpp.
-# The default event loop used on windows was changed from SelectorEventLoop
-# to ProactorEventLoop which is still not supported by aioxmpp.
-def get_event_loop() -> asyncio.SelectorEventLoop:
-    if sys.platform == 'win32':
-        policy = asyncio.get_event_loop_policy()
-        loop = policy._local._loop
-
-        if loop is None:
-            selector = selectors.SelectSelector()
-            loop = asyncio.SelectorEventLoop(selector)
-            asyncio.set_event_loop(loop)
-
-        elif isinstance(loop, asyncio.ProactorEventLoop):
-            raise RuntimeError('asyncio.ProactorEventLoop is not supported')
-
-    else:
-        loop = asyncio.get_event_loop()
-
-    return loop
-
-
 async def _start_client(client: 'Client', *,
                         shutdown_on_error: bool = True,
                         after: Optional[AnyCallableWithClient] = None
                         ) -> None:
-    loop = get_event_loop()
+    loop = asyncio.get_event_loop()
 
     if not isinstance(client, Client):
         raise TypeError('client must be an instance of fortnitepy.Client')
@@ -210,7 +187,7 @@ async def start_multiple(clients: List['Client'], *,
     AuthException
         An error occured when attempting to log in.
     """  # noqa
-    loop = get_event_loop()
+    loop = asyncio.get_event_loop()
 
     async def all_ready_callback_runner():
         tasks = [loop.create_task(client.wait_until_ready())
@@ -263,7 +240,7 @@ async def close_multiple(clients: List['Client']) -> None:
     clients: List[:class:`Client`]
         A list of the clients you wish to close.
     """
-    loop = get_event_loop()
+    loop = asyncio.get_event_loop()
 
     tasks = [loop.create_task(client.logout())
              for client in clients if not client._closing]
@@ -307,7 +284,7 @@ def run_multiple(clients: List['Client'], *,
     AuthException
         An error occured when attempting to log in.
     """  # noqa
-    loop = get_event_loop()
+    loop = asyncio.get_event_loop()
     _stopped = False
 
     def stopper(*args):
@@ -442,7 +419,7 @@ class Client:
                  cache_users: bool = True,
                  **kwargs: Any) -> None:
 
-        self.loop = loop or get_event_loop()
+        self.loop = loop or asyncio.get_event_loop()
         self.cache_users = cache_users
 
         self.status = kwargs.get('status', 'Battle Royale Lobby - {party_size} / {party_max_size}')  # noqa
@@ -2192,11 +2169,7 @@ class Client:
             if party_data['config']['joinability'] == 'INVITE_AND_FORMER':
                 raise Forbidden('You can\'t join a private party.')
 
-        party = ClientParty(self, party_data)
-        await party._update_members(party_data['members'])
-
         await self.user.party._leave()
-        self.user.set_party(party)
 
         future = asyncio.ensure_future(self.wait_for(
             'party_member_join',
@@ -2205,14 +2178,22 @@ class Client:
 
         try:
             await self.http.party_join_request(party_id)
-            await self.user.party.join_chat()
         except HTTPException as e:
+            if not future.cancelled():
+                future.cancel()
+
             await self._create_party()
 
             if e.message_code == ('errors.com.epicgames.social.'
                                   'party.party_join_forbidden'):
                 raise Forbidden('Client has no right to join this party.')
             e.reraise()
+
+        party_data = await self.http.party_lookup(party_id)
+        party = ClientParty(self, party_data)
+        self.user.set_party(party)
+        asyncio.ensure_future(party.join_chat(), loop=self.loop)
+        await party._update_members(party_data['members'])
 
         try:
             await future