+++ fortnitepy/http.py
@@ -3,7 +3,7 @@
 """
 MIT License
 
-Copyright (c) 2019 Terbau
+Copyright (c) 2019-2020 Terbau
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -748,12 +748,13 @@ class HTTPClient:
         )
         return await self.get(r)
 
-    async def account_get_by_user_id(self, user_id: str) -> dict:
+    async def account_get_by_user_id(self, user_id: str, *,
+                                     auth: Optional[str] = None) -> dict:
         r = AccountPublicService(
             '/account/api/public/account/{user_id}',
             user_id=user_id
         )
-        return await self.get(r)
+        return await self.get(r, auth=auth)
 
     async def account_get_by_email(self, email: str) -> dict:
         r = AccountPublicService(